Get rid of credential check.
The credential check was trying to list all buckets. With AWS
permissioning, you may be able to list/delete things in a specific
bucket, but not have access to the list of all buckets. So the
credential check was failing, even though the script would have worked.

There probably should be a better credential check, but we don't need it
for now.